issue=#347 add data & snapshot case
