+++ ft_test.sh
@@ -9,9 +9,12 @@ mkdir -p tmp/data
 cp -r test/testcase tmp/bin
 
 cd tmp/bin/
+sh kill_tera.sh
 sh launch_tera.sh
+sleep 3
+
 export PYTHONPATH=$PYTHONPATH:../../thirdparty/include/; export PATH=$PATH:../../thirdparty/bin/ 
-cd testcase
-nosetests -s -v > ../../log/test.log
+nosetests -s -v > ../log/test.log
+
 sh kill_tera.sh
