+++ internal_build.sh
@@ -12,6 +12,7 @@ DEPS_PREFIX=`pwd`/thirdparty
 DEPS_CONFIG="--prefix=${DEPS_PREFIX} --disable-shared --with-pic"
 
 export PATH=${DEPS_PREFIX}/bin:$PATH
+
 mkdir -p ${DEPS_SOURCE} ${DEPS_PREFIX}
 git clone --depth=1 http://gitlab.baidu.com/baidups/third.git ${DEPS_SOURCE}
 git clone --depth=1 http://gitlab.baidu.com/baidups/sofa-pbrpc.git ${DEPS_SOURCE}/sofa-pbrpc
@@ -143,3 +144,4 @@ sed -i 's/^INS_PREFIX=.*/INS_PREFIX=.\/thirdparty/' depends.mk
 ########################################
 
 make -j4
+
