+++ test/testcase/test_put_get.py
@@ -4,10 +4,6 @@ Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file.
 """
 
-import nose.tools
-import subprocess
-import time
-import os
 import common
 
 def setUp():
@@ -17,12 +13,10 @@ def setUp():
     
     common.print_debug_msg(0, "start master, ts1, ts2, ts3, and status is ok")
     common.print_debug_msg(1, "create test_table001 and test_table002(kv)")
-    cmd = "cd ../; ./teracli createbyfile testcase/data/create_table_schema; \
-           cd testcase/"
+    cmd = "./teracli createbyfile testcase/data/create_table_schema"
     common.exe_and_check_res(cmd)
 
-    cmd = "cd ../; ./teracli create 'table_test002 <storage=flash, splitsize=2048, mergesize=128>'; \
-           cd testcase/"
+    cmd = "./teracli create 'table_test002 <storage=flash, splitsize=2048, mergesize=128>'"
     common.exe_and_check_res(cmd)
 
 
@@ -33,11 +27,10 @@ def test_put_table():
 
     common.print_debug_msg(2, "put one data to table_test001 and table_test002")
     
-    cmd = "cd ../; ./teracli put table_test001 test001key update_flag:test001q test001v; \
-           cd testcase/"
+    cmd = "./teracli put table_test001 test001key update_flag:test001q test001v"
     common.exe_and_check_res(cmd)
 
-    cmd = "cd ../; ./teracli put table_test002 test002key test002v; cd testcase/"
+    cmd = "./teracli put table_test002 test002key test002v"
     common.exe_and_check_res(cmd)
 
 
@@ -48,10 +41,10 @@ def test_get_table():
 
     common.print_debug_msg(3, "read data form table_test001 and table_test002")
 
-    cmd = "cd ../; ./teracli get table_test001 test001key update_flag:test001q; cd testcase/"
+    cmd = "./teracli get table_test001 test001key update_flag:test001q"
     common.exe_and_check_res(cmd)
 
-    cmd = "cd ../; ./teracli get table_test002 test002key; cd testcase/"
+    cmd = "./teracli get table_test002 test002key"
     common.exe_and_check_res(cmd)
